pyproject.toml
[tool.poetry]
name = "mcp-demo"
version = "0.1.0"
description = ""
authors = ["Scott Hallock <ZSH4006@conagra.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.82.0"
fastmcp = "^2.5.1"
rich = "^14.0.0"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
